+++ src/Screens/MentorReg.tsx
@@ -1,166 +1,176 @@
-import { Button, Checkbox, FormControlLabel, Typography } from '@mui/material'
-import axios from 'axios'
-import React, { useEffect, useState } from 'react'
-import cred from '../utils/creds.json'
-import { signOut } from "firebase/auth";
-import { auth } from "../utils/firebase";
+import { Button, Checkbox, FormControlLabel, Typography } from "@mui/material";
+import {useState } from "react";
 import styling from './MentorRegStyles'
+import MentorComForm from "./MentorComForm";
 
+const MentorReg = (props:any) => {
 
-const Register = () => {
-    const url = cred.api_url;
-    const usertoken = localStorage.getItem("token");
-    const [callfunction, setcallfunction] = useState(false)
+  const [mentorpage1,setmentorpage1]= useState(true)
 
-    useEffect(() => {
-        axios.get(`${url}/university`, {
-            headers: {
-                'Authorization': `Bearer ${usertoken}`,
-            },
-            data: {
-                curredu: curredu
-            }
-        })
-            .then(response => {
+  //forms states
+  const [uni, setuni] = useState("");
+  const [course, setcourse] = useState("");
+  const [degree, setdegree] = useState("");
+  const [study, setstudy] = useState("");
+  const [relation, setrelation] = useState("");
+  const [uniemail, setuniemail] = useState("");
+  const [grad, setgrad] = useState("");
+  const [country, setcountry] = useState("");
+  const [previous, setprevious] = useState("");
+  const [specialization, setspecialization] = useState("");
+  const [campusjob, setcampusjob] = useState(false);
+  const [scholarship, setscholarship] = useState(false);
+  const [place, setplace] = useState("");
+  const [languages, setlanguages] = useState([]);
+  const [bio, setbio] = useState("");
+  const [gpa, setgpa] = useState("");
+  const [gmat, setgmat] = useState("");
+  const [sat, setsat] = useState("");
 
+  const classes=styling()
 
-            })
-            .catch(function( error ) {
-                if(error.response.status === 401) {
-                    signOut(auth).then()
-                }
-            })
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [callfunction])
+  console.log(props?.username);
+  console.log(props?.data);
+  
+  const handlecampus = () => {
+    (campusjob) ?
+      setcampusjob(false)
+      :
+      setcampusjob(true)
+  }
+  const handlescholarship = () => {
+    (scholarship) ?
+      setscholarship(false)
+      :
+      setscholarship(true)
+  }
 
+  return (
+    (mentorpage1)?
+    <div style={{ overflowY: "auto" }} className={classes.box}>
+    <Typography variant="h4" style={{ color: "#6E3CBC", fontWeight: "700", margin: "10px 0px" }}
+    >Mentor sign up</Typography>
+    <Typography variant="h6" style={{ marginBottom: "16px" }}>please enter
+      your details below to sign up as a mentor</Typography>
+    <div className={classes.innerbox}>
+      <label htmlFor="curredu" className={classes.title}> UNIVERSITY</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setuni(event.target.value)
+      }} value={uni} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> COURSE NAME</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setcourse(event.target.value)
+      }} value={course} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> DEGREE NAME</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setdegree(event.target.value)
+      }} value={degree} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> AREA OF STUDY</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setstudy(event.target.value)
+      }} value={study} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> UNIVERSITY RELATION</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setrelation(event.target.value)
+      }} value={relation} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> UNIVERSITY EMAIL ID (if current student)</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setuniemail(event.target.value)
+      }} value={uniemail} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> YEAR OF GRADUATION</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setgrad(event.target.value)
+      }} value={grad} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> COUNTRY OF STUDY</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setcountry(event.target.value)
+      }} value={country} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> PREVIOUS EDUCATIONAL INSTITUTE</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setprevious(event.target.value)
+      }} value={previous} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> SPECIALIZATION</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setspecialization(event.target.value)
+      }} value={specialization} required={true} />
+      <br />
+      <FormControlLabel
+        control={<Checkbox checked={campusjob} onChange={handlecampus} name="checkedA"
+          style={{ color: "#6E3CBC" }} />}
+        label="Ever had an on campus job ?"
+      />
+      <FormControlLabel
+        control={<Checkbox checked={scholarship} onChange={handlescholarship} name="checkedA"
+          style={{ color: "#6E3CBC" }} />}
+        label="Do you hold a scholarship ?"
+      />
+      <label htmlFor="curredu" className={classes.title}> PLACE OF STAY</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setplace(event.target.value)
+      }} value={place} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> PRIMARY AND SECONDARY LANGUAGES</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setlanguages((event: any) => { const language = (event.target.value).split(","); return language })
+      }} value={languages} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> BIO</label>
+      <textarea name="curredu" id="" cols={30} rows={10} className={classes.biggertext}
+        onChange={(event) => {
+          setbio(event.target.value)
+        }} value={bio} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> CURRENT GPA</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        // @ts-ignore
+        setgpa(event.target.value)
+      }} value={gpa} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> GMAT SCORE (FOR MASTER’S DEGREES)</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setgmat(event.target.value)
+      }} value={gmat} required={true} />
+      <br />
+      <label htmlFor="curredu" className={classes.title}> SAT SCORE (FOR BACHELOR’S DEGREES)</label>
+      <input type="text" name="curredu" className={classes.sminput} onChange={(event) => {
+        setsat(event.target.value)
+      }} value={sat} required={true} />
+      <br />
+      <Button variant="contained" className={classes.button} onClick={() => {
+        setmentorpage1(false)
+      }}>Submit</Button>
+    </div>
+  </div>:
+  <MentorComForm details={{
+    uni,
+    course,
+    degree,
+    study,
+    relation,
+    uniemail,
+    grad,
+    country,
+    previous,
+    specialization,
+    campusjob,
+    scholarship,
+    place,
+    languages,
+    bio,
+    gpa,
+    gmat,
+    sat
+  }}/>
+  )
+};
 
-    const [curredu, setcurredu] = useState("");
-    const [prevedu, setprevedu] = useState("");
-    const [campus, setcampus] = useState("");
-    const [pref1, setpref1] = useState("");
-    const [pref2, setpref2] = useState("");
-    const [pref3, setpref3] = useState("");
-    const [lang, setlang] = useState("");
-    const [bio, setbio] = useState("");
-    const [linkedin, setlinkedin] = useState("")
-    const [checked, setchecked] = useState(false)
-
-
-    const handleChange = () => {
-        ( checked ) ?
-            setchecked(false)
-            :
-            setchecked(true)
-    }
-
-    const classes = styling();
-    return (
-        <>
-            <div style={{ overflowY: "auto" }} className={classes.box}>
-                <Typography variant="h4" style={{ color: "#6E3CBC", fontWeight: "700", margin: "10px 0px" }}
-                >Mentor sign up</Typography>
-                <Typography variant="h6" style={{ marginBottom: "16px" }}>please enter
-                    your details below to sign up as a mentor</Typography>
-                <div className={classes.innerbox}>
-                    <label htmlFor="curredu" className={classes.title}> UNIVERSITY</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setcurredu(event.target.value)
-                    }} value={curredu} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> COURSE NAME</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setprevedu(event.target.value)
-                    }} value={prevedu} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> DEGREE NAME</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setcampus(event.target.value)
-                    }} value={campus} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> AREA OF STUDY</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setpref1(event.target.value)
-                    }} value={pref1} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> UNIVERSITY RELATION</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setpref2(event.target.value)
-                    }} value={pref2} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu"className={classes.title}> UNIVERSITY EMAIL ID (if current student)</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setpref3(event.target.value)
-                    }} value={pref3} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> YEAR OF GRADUATION</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> COUNTRY OF STUDY</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> PREVIOUS EDUCATIONAL INSTITUTE</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> SPECIALIZATION</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <FormControlLabel
-                        control={<Checkbox checked={checked} onChange={handleChange} name="checkedA"
-                                           style={{ color: "#6E3CBC" }}/>}
-                        label="Ever had an on campus job ?"
-                    />
-                    <FormControlLabel
-                        control={<Checkbox checked={checked} onChange={handleChange} name="checkedA"
-                                           style={{ color: "#6E3CBC" }}/>}
-                        label="Do you hold a scholarship ?"
-                    />
-                    <label htmlFor="curredu" className={classes.title}> PLACE OF STAY</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> PRIMARY AND SECONDARY LANGUAGES</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> BIO</label>
-                    <textarea name="curredu" id="" cols={30} rows={10} className={classes.biggertext}
-                              onChange={( event ) => {
-                                  setbio(event.target.value)
-                              }} value={bio} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> CURRENT GPA</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        // @ts-ignore
-                        setchecked(event.target.value)
-                    }} value={linkedin} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> GMAT SCORE (FOR MASTER’S DEGREES)</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <label htmlFor="curredu" className={classes.title}> SAT SCORE (FOR BACHELOR’S DEGREES)</label>
-                    <input type="text" name="curredu" className={classes.sminput} onChange={( event ) => {
-                        setlang(event.target.value)
-                    }} value={lang} required={true}/>
-                    <br/>
-                    <Button variant="contained" className={classes.button} onClick={() => {
-                        setcallfunction(true)
-                    }}>Submit</Button>
-                </div>
-            </div>
-        </>
-    )
-}
-
-export default Register
+export default MentorReg;
